Lab program 05
Design an XML document to store information about a student in a college affiliated to BU. The information must include USN, Name, Name of the College, Branch, Year of Joining, and e-mail id. Make up sample data for 3 students. Create a CSS style sheet and use it to display the document.